refactor: Soptamp와 관련된 Repository를 패키지로 묶음 (#332)
