+++ src/main/java/org/sopt/app/interfaces/postgres/soptamp_point/SoptampPointJpaRepository.java
@@ -1,10 +1,11 @@
-package org.sopt.app.interfaces.postgres;
+package org.sopt.app.interfaces.postgres.soptamp_point;
 
 import java.util.List;
 import java.util.Optional;
 import org.sopt.app.domain.entity.SoptampPoint;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SoptampPointRepository {
+public interface SoptampPointJpaRepository extends JpaRepository<SoptampPoint, Long> {
 
     List<SoptampPoint> findAllByGeneration(Long currentGeneration);
