+++ src/main/java/org/sopt/app/interfaces/postgres/soptamp_point/SoptampPointRepository.java
@@ -1,15 +1,19 @@
-package org.sopt.app.interfaces.postgres;
+package org.sopt.app.interfaces.postgres.soptamp_point;
 
 import java.util.List;
 import java.util.Optional;
 import org.sopt.app.domain.entity.SoptampPoint;
-import org.springframework.data.jpa.repository.JpaRepository;
+import org.sopt.app.domain.enums.Part;
 
-public interface SoptampPointJpaRepository extends JpaRepository<SoptampPoint, Long> {
+public interface SoptampPointRepository {
+
+    SoptampPoint save(SoptampPoint soptampPoint);
 
     List<SoptampPoint> findAllByGeneration(Long currentGeneration);
 
     Optional<SoptampPoint> findBySoptampUserIdAndGeneration(Long soptampUserId, Long currentGeneration);
 
     List<SoptampPoint> findAllBySoptampUserIdInAndGeneration(List<Long> soptampUserIdList, Long currentGeneration);
+
+    List<SoptampPoint> findAllBySoptampUserPartAndGeneration(Part part, Long currentGeneration);
 }
