+++ src/test/java/org/sopt/app/interfaces/postgres/SoptampPointRepositoryTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.sopt.app.common.config.QuerydslConfiguration;
 import org.sopt.app.domain.entity.SoptampPoint;
+import org.sopt.app.interfaces.postgres.soptamp_point.SoptampPointRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase.Replace;
